ISAICP-2603: Adding to the last large commit. When you change the idea of a functionality, make sure the pattern applies to the whole code. I changed the order of the foreach loops.
Before, we were looping first through the graphs and then through the entities.
Now we are looping first through the entities and then through the graphs.

Before we were returning in case a graph was valid within the first loop because we were satisfied by the results of one graph.
Now we should not return because otherwise we skip entities, NOT graphs. Add an if clause to continue instead.

I am never surpised my my ability to build patterns and my stupidity that leaves blank spots behind...